RAM amount can be <1G, so we use Double
Instead of Int

+++ Haskell/source/SourceModel.hs
@@ -92,7 +92,7 @@ data Static = Static {
 data InstanceType = InstanceType {
   typeID :: String
   , typeCPUs :: Int
-  , typeRAM :: Int
+  , typeRAM :: Double
   , typeCost :: Double
   } deriving (Show, Eq)
